Fixed size
Tightened up the code and switched from a for loop to a while loop for a little more control. The naming section was commented out as it will not be in the final program and will be replaced by a function that will automatically name the map. The dimensions of the map are now correct.