more cleanup and additions
Continuing to cleanup the code and filter it down to only the functional parts. It seems there may be an issue in bash where the screen shows 1 more line is available than there actually is so the maps y axis can never be 43 when drawn. I've started implementing map functions that will draw specific types of maps but I'm a long way from knowing exactly how to do this right now.