Many Changes
Changed the way the map is built and displayed for later convienence when adding things to the map. Getting ready to bring lungon.sh into the game as a control interface which will allow player interaction with the game. Probably further updates to the way the map is done to come so that player interaction is smoother and the game doesn't have to work so hard.